refactor: fitness 상위 패키지에 exercise 추가
- fitness 상위 패키지에 exercise 추가
