+++ src/main/java/com/flab/eattofit/exercise/application/fitness/dto/FitnessCreateRequest.java
@@ -1,0 +1,9 @@
+package com.flab.eattofit.exercise.application.fitness.dto;
+
+import jakarta.validation.constraints.NotEmpty;
+
+public record FitnessCreateRequest(
+        @NotEmpty(message = "등록할 피트니스 이름이 필요합니다.")
+        String name
+) {
+}
