+++ src/main/java/com/flab/eattofit/exercise/domain/fitness/Fitness.java
@@ -1,0 +1,35 @@
+package com.flab.eattofit.exercise.domain.fitness;
+
+import jakarta.persistence.Column;
+import jakarta.persistence.Entity;
+import jakarta.persistence.GeneratedValue;
+import jakarta.persistence.GenerationType;
+import jakarta.persistence.Id;
+import lombok.AccessLevel;
+import lombok.AllArgsConstructor;
+import lombok.Builder;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+
+@Getter
+@AllArgsConstructor(access = AccessLevel.PRIVATE)
+@NoArgsConstructor(access = AccessLevel.PROTECTED)
+@Builder
+@Entity
+public class Fitness {
+
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    private Long id;
+
+    @Column(unique = true)
+    private String name;
+
+    private Fitness(final String name) {
+        this.name = name;
+    }
+
+    public static Fitness from(final String name) {
+        return new Fitness(name);
+    }
+}
