+++ src/main/java/com/flab/eattofit/exercise/domain/fitness/FitnessRepository.java
@@ -1,0 +1,7 @@
+package com.flab.eattofit.exercise.domain.fitness;
+
+public interface FitnessRepository {
+
+    Fitness save(Fitness fitness);
+    boolean existsByName(String name);
+}
