+++ src/main/java/com/flab/eattofit/exercise/ui/fitness/dto/FitnessCreateResponse.java
@@ -1,0 +1,6 @@
+package com.flab.eattofit.exercise.ui.fitness.dto;
+
+public record FitnessCreateResponse(
+        String name
+) {
+}
