--- src/main/java/com/flab/eattofit/fitness/application/dto/FitnessCreateRequest.java
@@ -1,9 +1,0 @@
-package com.flab.eattofit.fitness.application.dto;
-
-import jakarta.validation.constraints.NotEmpty;
-
-public record FitnessCreateRequest(
-        @NotEmpty(message = "등록할 피트니스 이름이 필요합니다.")
-        String name
-) {
-}
