--- src/main/java/com/flab/eattofit/fitness/domain/FitnessRepository.java
@@ -1,7 +1,0 @@
-package com.flab.eattofit.fitness.domain;
-
-public interface FitnessRepository {
-
-    Fitness save(Fitness fitness);
-    boolean existsByName(String name);
-}
