--- src/main/java/com/flab/eattofit/fitness/ui/dto/FitnessCreateResponse.java
@@ -1,6 +1,0 @@
-package com.flab.eattofit.fitness.ui.dto;
-
-public record FitnessCreateResponse(
-        String name
-) {
-}
