--- src/test/java/com/flab/eattofit/fitness/fixture/FitnessCreateRequestFixture.java
@@ -1,14 +1,0 @@
-package com.flab.eattofit.fitness.fixture;
-
-import com.flab.eattofit.fitness.application.dto.FitnessCreateRequest;
-import org.junit.jupiter.api.DisplayNameGeneration;
-import org.junit.jupiter.api.DisplayNameGenerator;
-
-@DisplayNameGeneration(DisplayNameGenerator.ReplaceUnderscores.class)
-@SuppressWarnings("NonAsciiCharacters")
-public class FitnessCreateRequestFixture {
-
-    public static FitnessCreateRequest 피트니스_등록_요청서() {
-        return new FitnessCreateRequest("덤벨프레스");
-    }
-}
