--- src/test/java/com/flab/eattofit/fitness/fixture/FitnessFixture.java
@@ -1,17 +1,0 @@
-package com.flab.eattofit.fitness.fixture;
-
-import com.flab.eattofit.fitness.domain.Fitness;
-import org.junit.jupiter.api.DisplayNameGeneration;
-import org.junit.jupiter.api.DisplayNameGenerator;
-
-@DisplayNameGeneration(DisplayNameGenerator.ReplaceUnderscores.class)
-@SuppressWarnings("NonAsciiCharacters")
-public class FitnessFixture {
-
-    public static Fitness 덤벨프레스_id있음() {
-        return Fitness.builder()
-                .id(1L)
-                .name("덤벨프레스")
-                .build();
-    }
-}
