+++ src/test/java/com/flab/eattofit/helper/MockBeanInjection.java
@@ -2,8 +2,8 @@ package com.flab.eattofit.helper;
 
 import com.amazonaws.services.s3.AmazonS3;
 import com.fasterxml.jackson.databind.json.JsonMapper;
-import com.flab.eattofit.fitness.application.FitnessService;
-import com.flab.eattofit.fitness.domain.FitnessRepository;
+import com.flab.eattofit.exercise.application.fitness.FitnessService;
+import com.flab.eattofit.exercise.domain.fitness.FitnessRepository;
 import com.flab.eattofit.member.application.auth.AuthService;
 import com.flab.eattofit.member.application.auth.OAuthManager;
 import com.flab.eattofit.member.domain.auth.RefreshTokenRepository;
